% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stan.R
\name{build_stan}
\alias{build_stan}
\title{Build a Stan model without locking R/RStudio}
\usage{
build_stan(file)
}
\arguments{
\item{file}{A character string specifying the path to a Stan file to build.}
}
\value{
an object of the class rstan::stan_model.
}
\description{
Build a Stan model without locking R/RStudio
}
