% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_stan.R
\name{clean_stan}
\alias{clean_stan}
\title{Delete temporary files created by a Stan session}
\usage{
clean_stan(report_all_clean = T)
}
\arguments{
\item{report_all_clean}{A logical value indicating whether this function should print an "All Clean!" message when complete.}
}
\value{
No value is returned
}
\description{
Delete temporary files created by a Stan session
}
