% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigStan.R
\name{cleanBigStan}
\alias{cleanBigStan}
\title{Delete temporary files created by a Big Stan session}
\usage{
cleanBigStan(reportAllClean = T)
}
\arguments{
\item{reportAllClean}{A logical value indicating whether this function should print an "All Clean!" message when complete.}
}
\value{
No value is returned
}
\description{
Delete temporary files created by a Big Stan session
}
