% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_stan.R
\name{collect_stan}
\alias{collect_stan}
\title{Collect the results from a Stan session}
\usage{
collect_stan()
}
\value{
An object of S4 class stanfit.
}
\description{
Collect the results from a Stan session
}
