% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigStan.R
\name{collectBigStan}
\alias{collectBigStan}
\title{Collect the results from a Big Stan session}
\usage{
collectBigStan()
}
\value{
An object of S4 class stanfit.
}
\description{
Collect the results from a Big Stan session
}
