% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_condition_post}
\alias{get_condition_post}
\title{Combine coefficient samples from a Stan model with the design matrix to compute posterior samples on the conditions of the design.}
\usage{
get_condition_post(from_stan, par, X = NULL, W = NULL, B = NULL,
  numeric_res = 0)
}
\arguments{
\item{from_stan}{A stanfit object from a call to rstan::stan() or rstan::sampling() containing posterior samples from the model.}

\item{par}{Character string naming the parameter on which to compute the summary table.}

\item{X}{Optional contrast matrix as yielded from get_contrast_matrix() indicating the contrasts passed as data when creating the above stanfit object.}

\item{W}{Optional contrast matrix as yielded from get_contrast_matrix() indicating the within-subjects contrasts passed as data when creating the above stanfit object. If present, X is ignored and B must be present.}

\item{B}{Optional contrast matrix as yielded from get_contrast_matrix() indicating the between-subjects contrasts passed as data when creating the above stanfit object. W must be present, else B is ignored.}

\item{numeric_res}{Integer value conveying the number of points within the range of a continuous variable to sample. If 0, the values from the data are used.}
}
\value{
If the `tibble` package is installed, a tibble; else, a data frame.
}
\description{
Combine coefficient samples from a Stan model with the design matrix to compute posterior samples on the conditions of the design.
}
