% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_subject_indices}
\alias{get_subject_indices}
\title{Get a table of start/end indices from subject label vector}
\usage{
get_subject_indices(x)
}
\arguments{
\item{x}{A vector of subject labels, sorted by said labels.}
}
\value{
A matrix with two columns (start & end) and as many rows as there are unique labels in the input
}
\description{
Given input of an ordered subject label vector, returns the indices into this
vector corresponding to the beginning and end of each subject.
}

