% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfhelmert_contrasts.R
\name{halfhelmert_contrasts}
\alias{halfhelmert_contrasts}
\title{Half-Helmert Contrasts}
\usage{
halfhelmert_contrasts(...)
}
\arguments{
\item{...}{Arguments passed to contr.helmert}
}
\value{
See Value in contr.helmert().
}
\description{
contr.helmert()/2.
}
