% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{names_from_WB}
\alias{names_from_WB}
\title{Figure out variable names given W & B contrast matrices}
\usage{
names_from_WB(W, B, reverse = F)
}
\arguments{
\item{W}{Optional contrast matrix as yielded from get_contrast_matrix() indicating the within-subjects contrasts passed as data when creating the above stanfit object. If present, X is ignored and B must be present.}

\item{B}{Optional contrast matrix as yielded from get_contrast_matrix() indicating the between-subjects contrasts passed as data when creating the above stanfit object. W must be present, else B is ignored.}

\item{reverse}{Logical indicating whether to reverse the order of the names.}
}
\value{
A vector of character strings indicating names of variables.
}
\description{
Figure out variable names given W & B contrast matrices
}

