% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{stan_summary}
\alias{stan_summary}
\title{Compute a summary table for the posterior samples for a parameter from a Stan model}
\usage{
stan_summary(from_stan, par, probs = c(0.5, 0.025, 0.975), digits = 2,
  X = NULL, W = NULL, B = NULL, is_cor = F, return_array = F)
}
\arguments{
\item{from_stan}{A stanfit object from a call to rstan::stan() or rstan::sampling() containing posterior samples from the model.}

\item{par}{Character string naming the parameter on which to compute the summary table.}

\item{probs}{Vector of probabilities to show as columns in the summary table.}

\item{digits}{Integer value indicating number of significant digits to show when printing the summary table.}

\item{X}{Optional contrast matrix as yielded from get_contrast_matrix() indicating the contrasts passed as data when creating the above stanfit object.}

\item{W}{Optional contrast matrix as yielded from get_contrast_matrix() indicating the within-subjects contrasts passed as data when creating the above stanfit object. If present, X is ignored and B must be present.}

\item{B}{Optional contrast matrix as yielded from get_contrast_matrix() indicating the between-subjects contrasts passed as data when creating the above stanfit object. W must be present, else B is ignored.}

\item{is_cor}{A logical indicating whether the parameter is a correlation matrix. If TRUE, returns only the upper-diagonal entries.}

\item{return_array}{A logical indicating whether to return the summary table as an array. Default is to print but not return anything.}
}
\value{
Either nothing (merely prints to screen) if return_array is FALSE, or the array corresponding to the summary table.
}
\description{
Compute a summary table for the posterior samples for a parameter from a Stan model
}
