% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_stan.R
\name{start_stan}
\alias{start_stan}
\title{Start a Stan session}
\usage{
start_stan(data, mod, cores = parallel::detectCores()/2,
  chains_per_core = 1, seed_start = 1, ...)
}
\arguments{
\item{data}{A list of objects expected by the model as data.}

\item{mod}{A stan model created by rstan::stan_model().}

\item{cores}{An integer value specifying the number of cores to run in parallel.}

\item{chains_per_core}{An integer value specifying the number of chain to run sequentially on each core}

\item{seed_start}{An integer specifying the seed to use for the first core. All subsequent cores will use seed values that increment seed_start.}

\item{...}{Additional named arguments passed to rstan::sampling().}
}
\value{
No value is returned.
}
\description{
Start a Stan session
}
