% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigStan.R
\name{startBigStan}
\alias{startBigStan}
\title{Start a Big Stan session}
\usage{
startBigStan(stanMod, stanData, iter = 2000,
  cores = parallel::detectCores(), stanArgs = NULL, seedStart = 1,
  warmup = NULL)
}
\arguments{
\item{stanMod}{A stan model created by rstan::stan_model().}

\item{stanData}{A list of objects expected by the model as data.}

\item{iter}{An integer value specifying the number of iterations to run per core.}

\item{cores}{An integer value specifying the number of cores to run in parallel.}

\item{stanArgs}{A string containing code to append inside the call to rstan::sampling() to specify non-default values for its arguments. For example, the string "pars='mu',control=list(adapt_delta=.9)" would only keep samples for the "mu" parameter and set the adapt_delta value to .9.}

\item{seedStart}{An integer specifying the seed to use for the first core. All subsequent cores will use seed values that increment seedStart.}

\item{warmup}{An integer value, smaller than `iter` that specifies the number of iterations to run as warmup.}
}
\value{
No value is returned.
}
\description{
Start a Big Stan session
}
