% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigStan.R
\name{watchBigStan}
\alias{watchBigStan}
\title{Watch the progress of a Big Stan session}
\usage{
watchBigStan(updateInterval = 1, one_line_per_chain = TRUE, spacing = 3)
}
\arguments{
\item{updateInterval}{Number of seconds to wait between updates.}

\item{one_line_per_chain}{A logical value specifying whether the progress for each chain should be printed on seperate lines.}

\item{spacing}{An integer value specifying the number of extra spaces to add to the end of each chain's progress string. Can be used to fix misalignment when one_line_per_chain is TRUE.}
}
\value{
No value is returned.
}
\description{
Watch the progress of a Big Stan session
}
