% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfsum_contrasts.R
\name{halfsum_contrasts}
\alias{halfsum_contrasts}
\title{Half-sum Contrasts}
\usage{
halfsum_contrasts(...)
}
\arguments{
\item{...}{Arguments passed to contr.sum}
}
\value{
See Value in contr.sum().
}
\description{
Returns contrasts whereby coeffients will be equal to the difference between
levels (in the case of 2-level factors). Simply contr.sum()/2.
}
