% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_stan.R
\name{start_stan}
\alias{start_stan}
\title{Start a Stan session}
\usage{
start_stan(data, mod, iter = 2000, cores = parallel::detectCores(),
  args = NULL, seed_start = 1, warmup = NULL)
}
\arguments{
\item{data}{A list of objects expected by the model as data.}

\item{mod}{A stan model created by rstan::stan_model().}

\item{iter}{An integer value specifying the number of iterations to run per core.}

\item{cores}{An integer value specifying the number of cores to run in parallel.}

\item{args}{A string containing code to append inside the call to rstan::sampling() to specify non-default values for its arguments. For example, the string "pars='mu',control=list(adapt_delta=.9)" would only keep samples for the "mu" parameter and set the adapt_delta value to .9.}

\item{seed_start}{An integer specifying the seed to use for the first core. All subsequent cores will use seed values that increment seed_start.}

\item{warmup}{An integer value, smaller than `iter` that specifies the number of iterations to run as warmup.}
}
\value{
No value is returned.
}
\description{
Start a Stan session
}
